frontend/functions/requirements.txt
firebase-functions==1.1.0
firebase-admin==6.2.0
google-generativeai>=0.3.1
flask>=2.0.0